Extracted VideoGrid into its own component.
